delete tab full functionality, SwiftUI refactor
