+++ MuscleMemory/Assets.xcassets/checkmark.colorset/Contents.json
