+++ MuscleMemory/Assets.xcassets/menuButton.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "lightbulb 1.png",
+      "filename" : "Component 4.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
@@ -12,12 +12,12 @@
           "value" : "dark"
         }
       ],
-      "filename" : "lightbulb 2.png",
+      "filename" : "Component 7.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "lightbulb.png",
+      "filename" : "Component 5.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
@@ -28,12 +28,12 @@
           "value" : "dark"
         }
       ],
-      "filename" : "lightbulb 3.png",
+      "filename" : "Component 8.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "menuButton.png",
+      "filename" : "Component 6.png",
       "idiom" : "universal",
       "scale" : "3x"
     },
@@ -44,7 +44,7 @@
           "value" : "dark"
         }
       ],
-      "filename" : "lightbulb 4.png",
+      "filename" : "Component 9.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
